Split collection router unit tests into separate files
